+++ example/lib/page/dialog_page.dart
@@ -21,6 +21,7 @@ class _DialogPageState extends State<DialogPage>
     super.initState();
     Future.microtask(() async {
       await showDialog<bool>(
+          useRootNavigator: false,
           builder: (BuildContext context) {
             return AlertDialog(
               shape: const RoundedRectangleBorder(),
@@ -34,6 +35,33 @@ class _DialogPageState extends State<DialogPage>
                       Navigator.of(context).pop(false);
                     },
                     child: const Text('关闭')),
+                InkWell(
+                    child: const Text('topPageRouteName'),
+                    onTap: () {
+                      print(
+                          'topPageRouteName=${FusionNavigator.topPageRouteName}');
+                    }),
+                InkWell(
+                    child: const Text('topRouteName'),
+                    onTap: () {
+                      print('topRouteName=${FusionNavigator.topRouteName}');
+                    }),
+                InkWell(
+                    child: const Text('push(flutter) /lifecycle'),
+                    onTap: () async {
+                      FusionNavigator.push(
+                        '/lifecycle',
+                        routeType: FusionRouteType.flutter,
+                      );
+                    }),
+                InkWell(
+                    child: const Text('push(flutter) /index'),
+                    onTap: () async {
+                      FusionNavigator.push(
+                        '/index',
+                        routeType: FusionRouteType.flutter,
+                      );
+                    }),
               ],
             );
           },